move TrackSettings Radio buttons into its own component
