+++ src/content/app/genome-browser/components/track-settings-panel/track-settings-views/RegularTrackSettings.tsx
@@ -15,13 +15,11 @@
  */
 
 import React from 'react';
+import { useAppSelector, type RootState } from 'src/store';
 
 import SlideToggle from 'src/shared/components/slide-toggle/SlideToggle';
-import RadioGroup, {
-  type RadioOptions
-} from 'src/shared/components/radio-group/RadioGroup';
+import TrackSettingsRadioOptions from './components/track-settings-radio-options/TrackSettingsRadioOptions';
 
-import { useAppSelector, type RootState } from 'src/store';
 import {
   getApplyToAllSettings,
   getBrowserSelectedCog,
@@ -42,21 +40,11 @@ export const RegularTrackSettings = () => {
   const { toggleApplyToAll } = useTrackSettings();
 
   const { updateTrackName } = useTrackSettings();
-  const radioOptions: RadioOptions = [
-    {
-      value: 'this_track',
-      label: 'This track'
-    },
-    {
-      value: 'all_tracks',
-      label: 'All tracks'
-    }
-  ];
+
   return (
     <div className={styles.trackSettings}>
       <div className={styles.section}>
-        <RadioGroup
-          options={radioOptions}
+        <TrackSettingsRadioOptions
           onChange={toggleApplyToAll}
           selectedOption={shouldApplyToAll ? 'all_tracks' : 'this_track'}
         />
