+++ src/content/app/genome-browser/components/track-settings-panel/track-settings-views/components/track-settings-radio-options/TrackSettingsRadioOptions.tsx
@@ -1,0 +1,49 @@
+/**
+ * See the NOTICE file distributed with this work for additional information
+ * regarding copyright ownership.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React from 'react';
+import RadioGroup, {
+  OptionValue,
+  RadioOptions
+} from 'src/shared/components/radio-group/RadioGroup';
+
+type Props = {
+  selectedOption: 'all_tracks' | 'this_track';
+  onChange: (value: OptionValue) => void;
+};
+
+const TrackSettingsRadioOptions = (props: Props) => {
+  const radioOptions: RadioOptions = [
+    {
+      value: 'this_track',
+      label: 'This track'
+    },
+    {
+      value: 'all_tracks',
+      label: 'All tracks'
+    }
+  ];
+
+  return (
+    <RadioGroup
+      options={radioOptions}
+      onChange={props.onChange}
+      selectedOption={props.selectedOption}
+    />
+  );
+};
+
+export default TrackSettingsRadioOptions;
